# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server using Gorilla Mux with multiple APIs.",
        "title": "Example API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/oauth/google": {
            "get": {
                "description": "Exchanges the authorization code for an access token and retrieves user info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "Google Drive OAuth2 callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User info",
                        "schema": {
                            "$ref": "#/definitions/apis.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Missing code parameter",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to exchange token or retrieve user info",
                        "schema": {}
                    }
                }
            }
        },
        "/oauth/google": {
            "get": {
                "description": "Redirects the user to the Google OAuth2 authorization URL to obtain an access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "Redirect to Google OAuth2 authorization URL",
                "responses": {
                    "302": {
                        "description": "Redirect URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transfer": {
            "post": {
                "description": "Initiates the transfer of ownership for a Google Drive folder to a specified email address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "Transfer ownership of a Google Drive folder",
                "parameters": [
                    {
                        "description": "Request payload for transferring ownership",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ownership transfer initiated.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apis.TransferRequest": {
            "description": "Represents the request payload for transferring ownership.",
            "type": "object",
            "properties": {
                "folder_id": {
                    "description": "FolderID of the folder whose ownership is to be transferred\nRequired: true",
                    "type": "string",
                    "example": "1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgvE2upms"
                },
                "new_owner_email": {
                    "description": "Email of the new owner\nRequired: true",
                    "type": "string",
                    "example": "newowner@example.com"
                },
                "origin_email": {
                    "description": "Email of the current owner\nRequired: true",
                    "type": "string",
                    "example": "currnetowner@example.com"
                }
            }
        },
        "apis.UserInfo": {
            "description": "Represents the user info response.",
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email of the user\nRequired: true",
                    "type": "string"
                },
                "exp": {
                    "description": "Expiry time of the token\nRequired: true",
                    "type": "integer"
                },
                "picture": {
                    "description": "Picture of the user\nRequired: true",
                    "type": "string"
                }
            }
        }
    }
}